consolidating to one command line app
