+++ src/main/scala/com/josephpconley/swagger2postman/utils/HttpEnabled.scala
@@ -1,0 +1,6 @@
+package com.josephpconley.swagger2postman.utils
+
+trait HttpEnabled {
+
+  def execute(url: String): String
+}
